Add files via upload
It takes a little skill to create a calculator with basic mathematical functionalities using CSS, HTML and JS. Calculators require an interface to perform basic functions such addition, subtraction, division, multiplication. It will need a display screen to display the users input and give the results. The grid system in css is for the alignments of buttons on the calculator. The additional tools you need to build a fully functional calculator include eventListeners, if-else statements, operators, loops and so on